Work has begun on the "contact.html" file.
html - added the main "contact" block. In the course of work, tag <body> got a new class, for more convenient work with the size of the page. Classes were added for the block "contact", tags and content. In the process, the tags in the classes were slightly reworked.

css - "contact" block and its classes were added to style.scss/css file, page size was changed a little.

fix - the class of tag <body> in file "about.html" was changed.

@AidenOrtiz